requirements.txt
anyio==4.2.0
asgiref==3.7.2
astor==0.8.1
attrdict==2.0.1
Babel==2.14.0
bce-python-sdk==0.8.97
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cssselect==1.2.0
cssutils==2.9.0
cycler==0.12.1
Cython==3.0.6
decorator==5.1.1
Django==5.0
et-xmlfile==1.1.0
Flask==3.0.0
flask-babel==4.0.0
fonttools==4.46.0
future==0.18.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
image==1.5.33
imageio==2.33.1
imgaug==0.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
lmdb==1.4.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
networkx==3.2.1
numpy==1.26.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
paddleocr==2.6.0.1
paddlepaddle==2.5.2
pandas==2.1.4
Pillow==10.1.0
premailer==3.10.0
protobuf==3.20.2
psutil==5.9.7
pyclipper==1.3.0.post5
pycryptodome==3.19.0
Pygments==2.17.2
PyMuPDF==1.21.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
rapidfuzz==3.5.2
rarfile==4.1
requests==2.31.0
rich==13.7.0
scikit-image==0.22.0
scipy==1.11.4
Shapely @ file:///C:/Users/Tessiess/Downloads/Shapely-1.8.2-cp311-cp311-win_amd64.whl#sha256=1b2a4b58a437c342a9de40e062e80118d2d74a873e9e29f839f7310b3066f81e
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
thread==0.1.3
tifffile==2023.12.9
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
visualdl==2.5.3
Werkzeug==3.0.1
